Add TaskHistory for storage of sync metrics
This adds a new table for storing sync metrics and a quartz task for
cleaning up old entries in the table.

For now this only stores sync metrics, but it seems useful to also
store information about when pulses are sent and other background
tasks that would be useful to have additional information on when
users have issues.

Fixes #7611